requirements.txt
streamlit==1.37.1
python-dotenv==0.21.0
langchain==0.3.20
langchain-community==0.3.19
langchain-openai==0.3.8
faiss-cpu==1.10.0